+++ .travis.yml
@@ -3,6 +3,7 @@ language: python
 env:
   - DJANGO=1.7
   - DJANGO=1.8
+  - DJANGO=1.9
 python:
   - 2.7
   - 3.4
@@ -15,3 +16,6 @@ install:
   - pip install django-dirtyfield
   - pip install diff-match-patch==20121119
   - pip install six
+branches:
+  only:
+    - master
