+++ CHANGELOG.md
@@ -1,0 +1,10 @@
+Changelog
+=========
+
+0.4
+-----
+
+Changes:
+ - Django 1.9 support
+ - django-dirtyfield >= 0.9
+
