+++ MANIFEST.in
@@ -1,1 +1,2 @@
 graft djangohistory
+global-exclude *.pyc
