+++ djangohistory/__init__.py
@@ -1,2 +1,2 @@
 from __future__ import absolute_import
-from djangohistory.signals import *
+default_app_config = "djangohistory.apps.DjangoHistoryAppConfig"
