+++ djangohistory/apps.py
@@ -1,0 +1,8 @@
+from django.apps import AppConfig
+
+class DjangoHistoryAppConfig(AppConfig):
+    name = 'djangohistory'
+    def ready(self):
+        from djangohistory import signals
+
+
