+++ djangohistory/signals.py
@@ -1,3 +1,4 @@
+import django
 from django.conf import settings
 from django.db.models.signals import pre_delete, post_delete, post_init, post_save, m2m_changed, pre_save
 from django.db import models
@@ -5,12 +6,28 @@ from django.db import models
 from djangohistory.models import History
 from djangohistory.helpers import get_relation
 
+from djangodirtyfield.mixin import DirtyFieldMixin
+
 import six
 
 ACTION_MAP = {
 'post_add': 'm2m.add',
 'post_remove': 'm2m.remove',
 }
+
+def m2m_relations(instance):
+    fields = []
+    if django.VERSION[:2]>=(1, 9):
+        fields = [f for f in instance._meta.get_fields()
+                if f.is_relation and f.many_to_many]
+    else:
+        for field, model in instance._meta.get_m2m_with_model():
+            if isinstance(field, models.ManyToManyField):
+                fields.append(field)
+    return fields
+
+def is_df(instance):
+    return isinstance(instance, DirtyFieldMixin)
 
 def handle_m2m(sender, *args, **kwargs):
     """
@@ -21,11 +38,11 @@ def handle_m2m(sender, *args, **kwargs):
     """
     action = kwargs['action']
     instance = kwargs['instance']
-    if hasattr(instance, 'get_changes') and (action == 'post_add' or action == 'post_remove'):
+    if is_df(instance) and (action == 'post_add' or action == 'post_remove'):
         pk_set = list(kwargs['pk_set'])
         relation_name = sender._meta.db_table.replace(sender._meta.app_label + '_' + instance.__class__.__name__.lower() + '_', '')
         for pk in pk_set:
-            relations = {k.name:k for k in instance.get_m2m_relations()}
+            relations = {k.name:k for k in m2m_relations(instance)}
             related_instance = get_relation(relations[relation_name]).objects.get(pk=pk)
             field = relations[relation_name]
             changes = {field.name: {'changed': [pk], 'changed_to_string': six.text_type(related_instance)}}
@@ -38,10 +55,10 @@ def handle_m2m(sender, *args, **kwargs):
 
 def handle_model(sender, *args, **kwargs):
     instance = kwargs['instance']
-    if hasattr(instance, 'get_changes'):
+    if is_df(instance):
         changes = instance.get_changes()
         if changes.get('id') and changes['id']['old'] is None:
-            changes = instance.get_field_values()
+            changes = instance.dirtyfield.get_field_values()
             default_values = {}
             for k,v in six.iteritems(changes):
                 default_values[k] = {}
@@ -53,8 +70,8 @@ def handle_model(sender, *args, **kwargs):
 
 def handle_delete(sender, *args, **kwargs):
     instance = kwargs['instance']
-    if hasattr(instance, 'get_changes'): # TODO: settings.TRACKED_APPS/MODELS?
-        changes = instance.get_changes(dirty_fields=instance.get_field_values())
+    if is_df(instance): # TODO: settings.TRACKED_APPS/MODELS?
+        changes = instance.get_changes(dirty_fields=instance.dirtyfield.get_field_values())
         for k,v in six.iteritems(changes):
             v['new'] = ''
         History.objects.add(
