+++ requirements.txt
@@ -1,3 +1,3 @@
-Django==1.7
-django-dirtyfield
+Django==1.9
+django-dirtyfield>=0.9
 diff-match-patch==20121119
